+++ .github/CODEOWNERS
@@ -19,8 +19,7 @@
 /src/platform_impl/linux/x11        @kchibisov @notgull
 
 # Web
-/src/platform/web.rs                @daxpedda
-/src/platform_impl/web              @daxpedda
+/winit-web                          @daxpedda
 
 # Windows
 /src/platform/windows.rs            @notgull
