+++ .github/workflows/ci.yml
@@ -186,6 +186,10 @@ jobs:
       if: contains(matrix.platform.target, 'redox')
       run: cargo test -p winit-orbital
 
+    - name: Test winit Web
+      if: contains(matrix.platform.target, 'wasm')
+      run: cargo test -p winit-web
+
     # Test only on Linux x86_64, so we avoid spending unnecessary CI hours.
     - name: Test dpi crate
       if: >
