+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["dpi", "winit-core", "winit-orbital"]
+members = ["dpi", "winit-*"]
 resolver = "2"
 
 [workspace.package]
@@ -14,6 +14,7 @@ rust-version = "1.80"
 winit = { path = "." }
 winit-core = { version = "0.0.0", path = "winit-core" }
 winit-orbital = { version = "0.0.0", path = "winit-orbital" }
+winit-web = { version = "0.0.0", path = "winit-web" }
 
 # Core dependencies.
 bitflags = "2"
@@ -101,17 +102,7 @@ categories = ["gui"]
 description = "Cross-platform window creation library."
 documentation = "https://docs.rs/winit"
 edition.workspace = true
-include = [
-    "/build.rs",
-    "/docs",
-    "/examples",
-    "/FEATURES.md",
-    "/LICENSE",
-    "/src",
-    "!/src/platform_impl/web/script",
-    "/src/platform_impl/web/script/**/*.min.js",
-    "/tests",
-]
+include = ["/build.rs", "/docs", "/examples", "/FEATURES.md", "/LICENSE", "/src", "/tests"]
 keywords = ["windowing"]
 license.workspace = true
 name = "winit"
@@ -401,69 +392,8 @@ xkbcommon-dl.workspace = true
 [target.'cfg(target_os = "redox")'.dependencies]
 winit-orbital.workspace = true
 
-# Web
 [target.'cfg(target_family = "wasm")'.dependencies]
-js-sys.workspace = true
-pin-project.workspace = true
-wasm-bindgen.workspace = true
-wasm-bindgen-futures.workspace = true
-web-time.workspace = true
-web_sys = { workspace = true, features = [
-    "AbortController",
-    "AbortSignal",
-    "Blob",
-    "BlobPropertyBag",
-    "console",
-    "CssStyleDeclaration",
-    "Document",
-    "DomException",
-    "DomRect",
-    "DomRectReadOnly",
-    "Element",
-    "Event",
-    "EventTarget",
-    "FocusEvent",
-    "HtmlCanvasElement",
-    "HtmlElement",
-    "HtmlHtmlElement",
-    "HtmlImageElement",
-    "ImageBitmap",
-    "ImageBitmapOptions",
-    "ImageBitmapRenderingContext",
-    "ImageData",
-    "IntersectionObserver",
-    "IntersectionObserverEntry",
-    "KeyboardEvent",
-    "MediaQueryList",
-    "MessageChannel",
-    "MessagePort",
-    "Navigator",
-    "Node",
-    "OrientationLockType",
-    "OrientationType",
-    "PageTransitionEvent",
-    "Permissions",
-    "PermissionState",
-    "PermissionStatus",
-    "PointerEvent",
-    "PremultiplyAlpha",
-    "ResizeObserver",
-    "ResizeObserverBoxOptions",
-    "ResizeObserverEntry",
-    "ResizeObserverOptions",
-    "ResizeObserverSize",
-    "Screen",
-    "ScreenOrientation",
-    "Url",
-    "VisibilityState",
-    "WheelEvent",
-    "Window",
-    "Worker",
-] }
-
-[target.'cfg(all(target_family = "wasm", target_feature = "atomics"))'.dependencies]
-atomic-waker.workspace = true
-concurrent-queue.workspace = true
+winit-web.workspace = true
 
 [target.'cfg(target_family = "wasm")'.dev-dependencies]
 console_error_panic_hook.workspace = true
