+++ winit-web/src/os_error_macro.rs
@@ -1,0 +1,6 @@
+#[allow(unused_macros)]
+macro_rules! os_error {
+    ($error:expr) => {{
+        winit_core::error::OsError::new(line!(), file!(), $error)
+    }};
+}
